modular tree datastructure layout
